Add files via upload
hi all! i added script to filter out tweets with certain flagged words which we can determine in the beginning of the script to optimize our results. It is not a SQL filter, so it isn't ideal, but it is at least functional. I will go into more detail tomorrow :)